Agregar modales de confirmación

+++ back/public/js/dashboard.js
@@ -22,47 +22,72 @@ if (tableBody) {
       //location.assign('/producto/modificar');
     };
   }
-
-
-
+  
   const arrayBtnBajar = tableBody.getElementsByClassName('bajar');
   for (let boton of arrayBtnBajar) {
     document.getElementById(boton.id).onclick = async () => {
-      try {
-        const id = boton.id.split("-")[2];
-        console.log(id);
-        const response = await fetch(`${serverUrl}/producto/activo/${id}`, {
-          method: "PUT"
-        });
-        if (response.ok) {
-          console.log(`Producto Nº ${id} dado de baja correctamente`);
-          location.reload();
-        } else {
-          console.log('Error');
+      confirmarBaja().then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            const id = boton.id.split("-")[2];
+            const response = await fetch(`${serverUrl}/producto/activo/${id}`, {
+              method: "PUT"
+            });
+            if (response.ok) {
+              location.reload();
+            } else {
+              console.log('Error');
+            }
+          } catch (error) {
+            console.log('Error:', error);
+          }
         }
-      } catch (error) {
-        console.log('Error:', error);
-      }
+      });
     };
   }
 
   const arrayBtnEliminar = tableBody.getElementsByClassName('eliminar');
   for (let boton of arrayBtnEliminar) {
     document.getElementById(boton.id).onclick = async () => {
-      try {
-        const id = boton.id.split("-")[2];
-        const response = await fetch(`${serverUrl}/producto/${id}`, {
-          method: "DELETE"
-          //TODO: baja logica
-        });
-        if (response.ok) {
-          location.reload();
-        } else {
-          console.log('Error');
+      confirmarEliminacion().then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            const id = boton.id.split("-")[2];
+            const response = await fetch(`${serverUrl}/producto/${id}`, {
+              method: "DELETE"
+              //TODO: baja logica
+            });
+            if (response.ok) {
+              location.reload();
+            } else {
+              console.log('Error');
+            }
+          } catch (error) {
+            console.log('Error:', error);
+          }
         }
-      } catch (error) {
-        console.log('Error:', error);
-      }
+      });
     };
   }
 }
+
+const confirmarBaja = () => Swal.fire({
+  title: "¿Está seguro que desea cambiar el estado del producto?",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "Si",
+  cancelButtonText: "No",
+});
+
+const confirmarEliminacion = () => Swal.fire({
+  title: "¿Está seguro que desea eliminar el producto de la base de datos?",
+  text: "¡Esta acción no se puede revertir!",
+  icon: "warning",
+  showCancelButton: true,
+  confirmButtonColor: "#3085d6",
+  cancelButtonColor: "#d33",
+  confirmButtonText: "Si",
+  cancelButtonText: "No",
+});
